refactor(linux): modularize shell and onboarding (#100)
Refactor the Linux companion shell into a thinner
coordinator by centralizing section metadata and controller
lookup in `shell_sections` and extracting major integrated
pages into dedicated section modules.

Move Dashboard, General, Config, Diagnostics, Environment,
and Debug out of `app_window.c` into their own
`SectionController`-backed implementations. Unify
active-section refresh through controller dispatch and
centralize controller teardown during window shutdown.

Split onboarding into a controller/lifecycle layer
(`onboarding.c`) and a view/rendering layer
(`onboarding_view.c`). Add a narrow test seam so onboarding
controller lifecycle behavior can be validated headlessly.

Wire the new files into Meson and add focused tests for
shell section registry behavior and onboarding controller
lifecycle.
