+++ apps/linux/src/section_config.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "section_controller.h"
+
+const SectionController* section_config_get(void);
