+++ apps/linux/src/section_dashboard.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "section_controller.h"
+
+const SectionController* section_dashboard_get(void);
