+++ apps/linux/src/section_debug.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "section_controller.h"
+
+const SectionController* section_debug_get(void);
