+++ apps/linux/src/section_diagnostics.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "section_controller.h"
+
+const SectionController* section_diagnostics_get(void);
