+++ apps/linux/src/section_environment.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "section_controller.h"
+
+const SectionController* section_environment_get(void);
