+++ apps/linux/src/section_general.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "section_controller.h"
+
+const SectionController* section_general_get(void);
