+++ apps/linux/src/section_instances.h
@@ -11,7 +11,3 @@
 #include "section_controller.h"
 
 const SectionController* section_instances_get(void);
-
-/* Local instance card refresh (no RPC; uses local state/health/systemd).
- * Called by app_window on every tick for cheap local data. */
-void section_instances_refresh_local(void);
